requirements-gpuhub.txt
numpy==1.26.4
opencv-python==4.8.1.78
imutils==0.5.4
scipy==1.15.3
dlib==19.24.6
matplotlib==3.10.8
pandas==2.3.3
seaborn==0.13.2
PyYAML==6.0.3
tqdm==4.67.3
Pillow==12.0.0
requests==2.32.5
fastapi==0.115.12
uvicorn[standard]==0.35.0
python-multipart==0.0.20
aiofiles==24.1.0
